requirements.txt
Django>=1.8,<1.10
django-haystack>=2.4,<2.5
whoosh>=2.7,<2.8
mistune==0.7.1
Pillow>=3.0,<3.1
django-infinite-scroll-pagination>=0.1.3,<0.2
unicode-slugify==0.1.3
pytz
